backend/pyproject.toml
[tool.poetry]
name = "\\beat_storage"
version = "0.1.0"
description = ""
authors = ["a.zinoviev <pandarado@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
django-environ = "^0.11.2"
fastapi = "^0.110.1"
loguru = "^0.7.2"
uvicorn = "^0.29.0"
arrow = "^1.3.0"
ldap3 = "^2.9.1"
gunicorn = "^21.2.0"
httpx = "^0.27.0"
psycopg2-binary = "^2.9.9"
django-admin-interface = "^0.28.6"
django-extensions = "^3.2.3"
django-admin-rangefilter = "^0.12.1"
django-settings-export = "^1.2.1"
pyjwt = "^2.8.0"
boto3 = "^1.35.41"
django-storages = "^1.14.4"
fastapi-pagination = "^0.12.31"
django-taggit = "^6.1.0"
django-modeladmin-reorder = "^0.3.1"
django-ordered-model = "^3.7.4"
django-parler = "^2.3"
django-taggit-labels = "^0.7.0"
django-taggit-anywhere = "^1.0.0"
django-taggit-helpers = "^0.1.4"
django-taggit-templatetags = "^0.2.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
